add: Status View 생성 및 VC로부터 분리
