+++ DustApp/DustApp/Controllers/DustViewController.swift
@@ -10,27 +10,16 @@ import UIKit
 import CoreLocation
 
 class DustViewController: UIViewController, CLLocationManagerDelegate {
-    @IBOutlet var gradientView: UIView!
-    @IBOutlet var statusImage: UIImageView!
-    @IBOutlet var statusLabel: UILabel!
-    @IBOutlet var statusValueLabel: UILabel!
-    @IBOutlet var measureTimeLabel: UILabel!
-    @IBOutlet var measurePlaceLabel: UILabel!
+    @IBOutlet var gradientView: StatusView!
     @IBOutlet var tableView: TimelineTableView!
     
     var locationManager: LocationManager!
     let dustNetworkManager = Dust24NetworkManager()
-//    var measuredValue:Int = 67
     
     override func viewDidLoad() {
         super.viewDidLoad()
         locationManager = LocationManager()
-        
         requestQueryWithCoordinate()
-        
-//        let state = measureDustGrade(measuredValue: measuredValue)
-//        inputUIValues(state: state)
-        
     }
     
     func requestQueryWithCoordinate() {
@@ -49,36 +38,5 @@ class DustViewController: UIViewController, CLLocationManagerDelegate {
             }
         }
     }
-    
-    func makeGradientView(gradientColor: [Any]) {
-        let gradient = CAGradientLayer()
-        gradient.frame = view.bounds
-        gradient.endPoint = CGPoint(x: 0.5, y: 1)
-        gradient.colors = gradientColor
-        gradientView.layer.insertSublayer(gradient, at: 0)
-    }
-    
-    func inputStatusImage(stateEmoji: UIImage) {
-        statusImage.image = stateEmoji
-    }
-    
-    func inputStatusLabel(state: String) {
-        statusLabel.text = state
-    }
-    
-    func inputStatusValueLabel(measuredValue: Int) {
-        statusValueLabel.text = "\(measuredValue)"
-    }
-    
-//    func measureDustGrade(measuredValue: Int) -> DustState {
-//        let grade = DustGrade(measuredValue)
-//        return grade.gradeDustState(measuredValue: measuredValue)
-//    }
-    
-    func inputUIValues(state: DustState) {
-        makeGradientView(gradientColor: state.gradientColor)
-        inputStatusImage(stateEmoji: state.stateEmoji)
-        inputStatusLabel(state: state.state)
-//        inputStatusValueLabel(measuredValue: measuredValue)
-    }
 }
+
